+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2021-01-17
+### Fixed
+- Fixed occasional IO errors due to incorrect HID report read buffer size.
+
 ## [0.3.0] - 2021-01-10
 ### Added
 - Implemented `std::fmt::Display` for the `Rail` enumeration.
@@ -29,7 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-01-07
 - Initial release
 
-[Unreleased]: https://github.com/newAM/corsairmi-rs/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/newAM/corsairmi-rs/compare/v0.3.1...HEAD
+[0.3.1]: https://github.com/newAM/corsairmi-rs/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/newAM/corsairmi-rs/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/newAM/corsairmi-rs/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/newAM/corsairmi-rs/releases/tag/v0.1.0